Update file structure/clean up
Extract replay and twitter code into their own folders

Add ESLint as a dev dependency
