+++ .eslintrc.json
@@ -1,0 +1,15 @@
+{
+    "env": {
+      "node": true,
+      "es2021": true
+    },
+    "extends": "eslint:recommended",
+    "parserOptions": {
+        "ecmaVersion": 12,
+        "sourceType": "module"
+    },
+    "globals": {
+    },
+    "rules": {
+    }
+}
