+++ .gitignore
@@ -54,6 +54,9 @@ typings/
 # Yarn Integrity file
 .yarn-integrity
 
+# MacOS
+.DS_Store
+
 # dotenv environment variables file
 .env
