+++ package.json
@@ -2,7 +2,7 @@
   "name": "Dustbot",
   "version": "1.4.1",
   "description": "Posts twitch streams to discord, replay data to the discord & twitter.",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -25,5 +25,8 @@
   },
   "engines": {
     "node": ">=16.9.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.33.0"
   }
 }
